[Decomp][ObjectMgr] Cohesion split: extract locale loaders to ObjectMgrLocales.cpp
